refactor: data structure and auth
